App review is back (#6367)
