+++ packages/smooth_app/lib/pages/user_management/login_page.dart
@@ -8,16 +8,12 @@ import 'package:smooth_app/data_models/login_result.dart';
 import 'package:smooth_app/data_models/preferences/user_preferences.dart';
 import 'package:smooth_app/data_models/user_management_provider.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
-import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
 import 'package:smooth_app/generic_lib/widgets/smooth_card.dart';
 import 'package:smooth_app/generic_lib/widgets/smooth_text_form_field.dart';
 import 'package:smooth_app/helpers/analytics_helper.dart';
 import 'package:smooth_app/helpers/app_helper.dart';
-import 'package:smooth_app/helpers/launch_url_helper.dart';
-import 'package:smooth_app/helpers/user_feedback_helper.dart';
 import 'package:smooth_app/pages/user_management/forgot_password_page.dart';
 import 'package:smooth_app/pages/user_management/sign_up_page.dart';
-import 'package:smooth_app/services/smooth_services.dart';
 import 'package:smooth_app/widgets/smooth_app_bar.dart';
 import 'package:smooth_app/widgets/smooth_scaffold.dart';
 
@@ -67,7 +63,6 @@ class _LoginPageState extends State<LoginPage> with TraceableClientMixin {
     }
 
     AnalyticsHelper.trackEvent(AnalyticsEvent.loginAction);
-    await _showInAppReviewIfNecessary(context);
     if (!context.mounted) {
       return;
     }
@@ -342,83 +337,5 @@ class _LoginPageState extends State<LoginPage> with TraceableClientMixin {
         ),
       ),
     );
-  }
-
-  Future<void> _showInAppReviewIfNecessary(BuildContext context) async {
-    final UserPreferences preferences = context.read<UserPreferences>();
-    if (preferences.inAppReviewAlreadyAsked) {
-      return;
-    }
-
-    final AppLocalizations appLocalizations = AppLocalizations.of(context);
-
-    final bool? enjoyingApp = await showDialog<bool>(
-      context: context,
-      builder: (BuildContext context) => SmoothAlertDialog(
-        title: appLocalizations.app_rating_dialog_title_enjoying_app,
-        body: const SizedBox.shrink(),
-        close: true,
-        actionsAxis: Axis.vertical,
-        positiveAction: SmoothActionButton(
-          text: appLocalizations.tagline_app_review_button_positive,
-          onPressed: () => Navigator.of(context).pop(true),
-        ),
-        negativeAction: SmoothActionButton(
-          text: appLocalizations.tagline_app_review_button_negative,
-          onPressed: () => Navigator.of(context).pop(false),
-        ),
-        neutralAction: SmoothActionButton(
-          text: appLocalizations.tagline_app_review_button_later,
-          onPressed: () => Navigator.of(context).pop(null),
-        ),
-      ),
-    );
-    if (enjoyingApp == null) {
-      return;
-    }
-    if (!context.mounted) {
-      return;
-    }
-    if (!enjoyingApp) {
-      await showDialog<void>(
-        context: context,
-        builder: (BuildContext context) => SmoothAlertDialog(
-          body: Text(appLocalizations.app_rating_dialog_title_not_enjoying_app),
-          positiveAction: SmoothActionButton(
-            text: appLocalizations.okay,
-            onPressed: () async {
-              final String formLink = UserFeedbackHelper.getFeedbackFormLink();
-              LaunchUrlHelper.launchURL(formLink);
-              Navigator.of(context).pop();
-            },
-          ),
-          negativeAction: SmoothActionButton(
-            text: appLocalizations.not_really,
-            onPressed: () => Navigator.of(context).pop(),
-          ),
-        ),
-      );
-      return;
-    }
-
-    final bool? userRatedApp = await showDialog<bool>(
-      context: context,
-      builder: (BuildContext context) => SmoothAlertDialog(
-        body: Text(appLocalizations.app_rating_dialog_title),
-        positiveAction: SmoothActionButton(
-          text: appLocalizations.app_rating_dialog_positive_action,
-          onPressed: () async => Navigator.of(context).pop(
-            await ApplicationStore.openAppReview(),
-          ),
-        ),
-        negativeAction: SmoothActionButton(
-          text: appLocalizations.ask_me_later_button_label,
-          onPressed: () => Navigator.of(context).pop(),
-        ),
-      ),
-    );
-    if (userRatedApp == true) {
-      await preferences.markInAppReviewAsShown();
-    }
   }
 }
